CMS-9963 Collapse to previous width instead of configured width
When enabling resizing for the second column of the content
perspective, the expand/collapse action should collapse back to
the previous width instead of the configured width, as the user
might have changed it.

To sync the UI during window resize, expanded units will be
expanded again, so only store the previous collapsed width when
unit is collapsed.